requirements.txt
Django==4.2.11   
channels==4.1.0
redis==4.5.5
psycopg2-binary==2.9.10
whitenoise==6.6.0
django-prometheus==2.3.1
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.1.0
bcrypt==4.2.1
gunicorn==21.2.0
python-dotenv==1.0.1
python-json-logger
uvicorn[standard]
websockets
wsproto
whitenoise
django_redis